+++ src/book_class.rb
