+++ src/nameable/base_decorator.rb
