+++ src/nameable/capitalize_decorator.rb
