+++ src/nameable/nameable_decorator.rb
