+++ src/nameable/trimmer_decorator.rb
