+++ src/rental_class.rb
